pyproject.toml
[build-system]
requires = ["setuptools>=65", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["templates/src"]
include = ["*"]

[project]
name = "vec-playbook"
version = "0.1.0"
description = "Starter templates that launch on Slurm via Hydra + Submitit"
authors = [{name = "Farnaz Kohankhaki", email = "farnaz.kohankhaki@vectorinstitute.ai"}]
requires-python = "==3.12.*"
dependencies = [
    "hydra-core>=1.3.2,<1.4",
    "hydra-submitit-launcher>=1.2.0,<1.5",
    "submitit>=1.5.0,<2.0",
    "torch>=2.0.0,<2.6.0",
    "transformers>=4.36.0,<4.52.0",
    "datasets>=2.16.0,<3.6.0",
    "accelerate>=0.26.0,<1.0.0",
    "pyarrow==16.1.0",
    "ruff>=0.1.0,<1.0.0",
    "pre-commit>=3.0.0,<4.0.0",
]

[tool.ruff]
include = ["*.py", "pyproject.toml", "*.ipynb"]
line-length = 88

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
docstring-code-format = true

[tool.ruff.lint]
select = ["A","B","COM","C4","RET","SIM","ICN","Q","RSE","D","E","F","I","W","N","ERA","PL"]
fixable = ["A","B","COM","C4","RET","SIM","ICN","Q","RSE","D","E","F","I","W","N","ERA","PL"]
ignore = ["B905","E501","D203","D213","PLR2004","PLR0913","COM812"]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["E402","F401","F403","F811"]

[tool.ruff.lint.pep8-naming]
ignore-names = ["X*", "setUp"]

[tool.ruff.lint.isort]
lines-after-imports = 2

[tool.ruff.lint.pydocstyle]
convention = "numpy"

[tool.ruff.lint.pycodestyle]
max-doc-length = 88
